requirements.txt
asgiref==3.8.1
dj-config-url==0.1.1
Django==5.1.6
djangorestframework==3.15.2
dotenv==0.9.9
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
